+++ projects/js-packages/critical-css-gen/tests/config/jest.config.js
@@ -1,9 +1,11 @@
+import coverageConfig from 'jetpack-js-tools/jest/config.coverage.js';
+
 export default {
+	...coverageConfig,
 	rootDir: '../',
 	testEnvironment: 'jest-environment-node',
 	testMatch: [ '**/?(*.)+(spec|test).js' ],
 	setupFilesAfterEnv: [ './config/jest-setup.js' ],
-	collectCoverageFrom: [ '../build-node/*.js' ],
 	testPathIgnorePatterns: [ '/node_modules/', 'config/jest-setup.js', 'build-node/*' ],
 	moduleDirectories: [ 'build-node', 'node_modules' ],
 };