+++ projects/packages/jetpack-mu-wpcom/src/features/help-center/class-help-center.php
@@ -285,6 +285,10 @@ class Help_Center {
 		$controller = new WP_REST_Help_Center_Support_Activity();
 		$controller->register_rest_route();
 
+		require_once __DIR__ . '/class-wp-rest-help-center-support-interactions.php';
+		$controller = new WP_REST_Help_Center_Support_Interactions();
+		$controller->register_rest_route();
+
 		require_once __DIR__ . '/class-wp-rest-help-center-user-fields.php';
 		$controller = new WP_REST_Help_Center_User_Fields();
 		$controller->register_rest_route();