+++ projects/js-packages/analytics/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Analytics package releases.
 
+## [0.1.32] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [0.1.31] - 2024-08-23
 ### Changed
 - Updated package dependencies. [#39004]
@@ -124,6 +128,7 @@
 ### Added
 - Initial release of jetpack-api package.
 
+[0.1.32]: https://github.com/Automattic/jetpack-analytics/compare/v0.1.31...v0.1.32
 [0.1.31]: https://github.com/Automattic/jetpack-analytics/compare/v0.1.30...v0.1.31
 [0.1.30]: https://github.com/Automattic/jetpack-analytics/compare/v0.1.29...v0.1.30
 [0.1.29]: https://github.com/Automattic/jetpack-analytics/compare/v0.1.28...v0.1.29