+++ projects/js-packages/config/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.26] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [0.1.25] - 2024-08-29
 ### Changed
 - Internal updates.
@@ -104,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - fixed and improved README
 
+[0.1.26]: https://github.com/Automattic/jetpack-config-js/compare/v0.1.25...v0.1.26
 [0.1.25]: https://github.com/Automattic/jetpack-config-js/compare/v0.1.24...v0.1.25
 [0.1.24]: https://github.com/Automattic/jetpack-config-js/compare/v0.1.23...v0.1.24
 [0.1.23]: https://github.com/Automattic/jetpack-config-js/compare/v0.1.22...v0.1.23