+++ projects/packages/schema/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [0.1.2] - 2024-10-10
 ### Fixed
 - Fix missing types in phpdoc comments. [#39648]
@@ -20,5 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#39004]
 
+[0.1.3]: https://github.com/Automattic/jetpack-schema/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Automattic/jetpack-schema/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/Automattic/jetpack-schema/compare/v0.1.0...v0.1.1