+++ projects/packages/wp-js-data-sync/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [0.5.1] - 2024-10-10
 ### Fixed
 - Fix missing types in phpdoc comments. [#39648]
@@ -89,6 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Jetpack Boost: Fix assoc_array schema validation. [#29611]
 
+[0.5.2]: https://github.com/Automattic/jetpack-wp-js-data-sync/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/Automattic/jetpack-wp-js-data-sync/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/Automattic/jetpack-wp-js-data-sync/compare/v0.4.5...v0.5.0
 [0.4.5]: https://github.com/Automattic/jetpack-wp-js-data-sync/compare/v0.4.4...v0.4.5